1
1

+++ 2019杭电分体式机器人1  一队/DRIVER/bottom_sensor.h
@@ -12,5 +12,5 @@ void ChangeCoordinate(void);
 void ROTATE(uint8_t Clockwise);//ClockwiseÎª1Ë³Ê±Õë
 void toFetch(uint8_t IsGet,uint8_t Floor);
 extern uint8_t LsRotate;
-extern float dis;
+extern double dis;
 #endif
